// src/apis-client/models/LeaseVisibilityInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aeler demo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LeaseVisibilityInput
 */
export interface LeaseVisibilityInput {
    /**
     * customerId
     * @type {string}
     * @memberof LeaseVisibilityInput
     */
    customerId: string;
    /**
     * start time
     * @type {Date}
     * @memberof LeaseVisibilityInput
     */
    startAt?: Date;
    /**
     * stop time
     * @type {Date}
     * @memberof LeaseVisibilityInput
     */
    stopAt?: Date;
    /**
     * flag
     * @type {boolean}
     * @memberof LeaseVisibilityInput
     */
    isSameDatesAsLease?: boolean;
    /**
     * flag
     * @type {boolean}
     * @memberof LeaseVisibilityInput
     */
    isSameDatesAsStatus?: boolean;
    /**
     * is it active or not
     * @type {boolean}
     * @memberof LeaseVisibilityInput
     */
    isActive?: boolean;
}

export function LeaseVisibilityInputFromJSON(json: any): LeaseVisibilityInput {
    return LeaseVisibilityInputFromJSONTyped(json, false);
}

export function LeaseVisibilityInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): LeaseVisibilityInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'customerId': json['customerId'],
        'startAt': !exists(json, 'startAt') ? undefined : (new Date(json['startAt'])),
        'stopAt': !exists(json, 'stopAt') ? undefined : (new Date(json['stopAt'])),
        'isSameDatesAsLease': !exists(json, 'isSameDatesAsLease') ? undefined : json['isSameDatesAsLease'],
        'isSameDatesAsStatus': !exists(json, 'isSameDatesAsStatus') ? undefined : json['isSameDatesAsStatus'],
        'isActive': !exists(json, 'isActive') ? undefined : json['isActive'],
    };
}

export function LeaseVisibilityInputToJSON(value?: LeaseVisibilityInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'customerId': value.customerId,
        'startAt': value.startAt === undefined ? undefined : (value.startAt.toISOString()),
        'stopAt': value.stopAt === undefined ? undefined : (value.stopAt.toISOString()),
        'isSameDatesAsLease': value.isSameDatesAsLease,
        'isSameDatesAsStatus': value.isSameDatesAsStatus,
        'isActive': value.isActive,
    };
}